move duplicate code to ingest pkg
