+++ packages/ingest/granule.js
@@ -596,6 +596,19 @@ function selector(type, protocol) {
 }
 
 /**
+ * Extract the granule ID from the a given s3 uri
+ *
+ * @param {string} uri - the s3 uri of the file
+ * @param {string} regex - the regex for extracting the ID
+ * @returns {string} the granule
+ */
+function getGranuleId(uri, regex) {
+  const match = path.basename(uri).match(regex);
+  if (match) return match[1];
+  return null;
+}
+
+/**
  * Gets metadata for a cmr xml file from s3
  *
  * @param {string} xmlFilePath - S3 URI to the xml metadata document
@@ -626,6 +639,37 @@ async function parseXmlString(xml) {
       return resolve(data);
     });
   });
+}
+
+/**
+ * returns a list of CMR xml files
+ *
+ * @param {Array} input - an array of s3 uris
+ * @param {string} granuleIdExtraction - a regex for extracting granule IDs
+ * @returns {Promise<Array>} promise resolves to an array of objects
+ * that includes CMR xmls uris and granuleIds
+ */
+async function getCmrFiles(input, granuleIdExtraction) {
+  const files = [];
+  const expectedFormat = /.*\.cmr\.xml$/;
+
+  input.forEach(async (filename) => {
+    if (filename && filename.match(expectedFormat)) {
+      const metadata = await getMetadata(filename);
+      const metadataObject = await parseXmlString(metadata);
+
+      const cmrFileObject = {
+        filename,
+        metadata,
+        metadataObject,
+        granuleId: getGranuleId(filename, granuleIdExtraction)
+      };
+
+      files.push(cmrFileObject);
+    }
+  });
+
+  return files;
 }
 
 async function postS3Object(destination, options) {
@@ -842,6 +886,8 @@ module.exports.S3Granule = S3Granule;
 module.exports.S3DiscoverGranules = S3DiscoverGranules;
 module.exports.SftpDiscoverGranules = SftpDiscoverGranules;
 module.exports.SftpGranule = SftpGranule;
+module.exports.getGranuleId = getGranuleId;
+module.exports.getCmrFiles = getCmrFiles;
 module.exports.getMetadata = getMetadata;
 module.exports.copyGranuleFile = copyGranuleFile;
 module.exports.moveGranuleFile = moveGranuleFile;
